+++ CMakeLists.txt
@@ -29,14 +29,17 @@ sec_generate_parameter_library(
   src/linear_feedback_controller.yaml # Path to input yaml file
 )
 add_dependencies(${PROJECT_NAME}_parameters ${PROJECT_NAME}_parameters_doc)
+
 # Joint estimator params
-sec_generate_parameter_library_markdown(joint_state_estimator_parameters_doc
-                                        src/joint_state_estimator.yaml)
+sec_generate_parameter_library_markdown(
+  joint_state_estimator_parameters_doc
+  src/joint_state_estimator/joint_state_estimator.yaml)
+
 sec_generate_parameter_library(
   joint_state_estimator_generated_parameters # Lib name
   joint_state_estimator_parameters # CMake target name for the parameter
                                    # library.
-  src/joint_state_estimator.yaml # Path to input yaml file
+  src/joint_state_estimator/joint_state_estimator.yaml # Path to input yaml file
 )
 add_dependencies(joint_state_estimator_parameters
                  joint_state_estimator_parameters_doc)
@@ -47,7 +50,7 @@ add_dependencies(joint_state_estimator_parameters
 set(${PROJECT_NAME}_headers
     include/${PROJECT_NAME}/averaging_filter.hpp
     include/${PROJECT_NAME}/averaging_filter.hxx
-    include/${PROJECT_NAME}/joint_state_estimator.hpp
+    include/joint_state_estimator/joint_state_estimator.hpp
     include/${PROJECT_NAME}/lf_controller.hpp
     include/${PROJECT_NAME}/linear_feedback_controller.hpp
     include/${PROJECT_NAME}/linear_feedback_controller_ros.hpp
@@ -55,7 +58,7 @@ set(${PROJECT_NAME}_headers
     include/${PROJECT_NAME}/pd_controller.hpp
     include/${PROJECT_NAME}/robot_model_builder.hpp)
 set(${PROJECT_NAME}_sources
-    src/joint_state_estimator.cpp #
+    src/joint_state_estimator/joint_state_estimator.cpp #
     src/lf_controller.cpp #
     src/linear_feedback_controller.cpp #
     src/linear_feedback_controller_ros.cpp #
